Draft: Cleanup in `IterationData`, `ControlResultsInfo`, and `ExperimentResult`

See merge request cpm/coincar/software/graph_based_planning!255

+++ hlc/optimizer/graph_search/tree/MonteCarloTree.m
@@ -103,7 +103,7 @@ classdef MonteCarloTree < Tree
         end
 
         function cost = get_cost(obj, i_node)
-            cost = obj.cost(1, i_node);
+            cost = double(obj.cost(1, i_node));
         end
 
     end
